v9 support added

+++ vttctl.sh
@@ -294,24 +294,28 @@ case "$1" in
         ;; 
   download)
         if [[ $2 =~ $REGEX_URL ]]; then 
-            DEST="FoundryVTT"
-            count=$(find $DEST/ -type d -name "[0-9][0-9].*[0-9][0-9][0-9]" | wc -l)
-            if [ $count -gt 9 ]; then
-                  echo "To many downloaded FoundryVTT binaries, please do $0 clean and remove at least 1 old version."
-                  break
-            fi
-            VERSION=$(echo "$2" | grep -oP "(?<=releases\/)\d+\.\d+")
-            TARGET="${DEST}/${VERSION}"
+            if [ MAJOR_VER -ge 9 ]; then
+                  DEST="FoundryVTT"
+                  count=$(find $DEST/ -type d -name "[0-9][0-9].*[0-9][0-9][0-9]" | wc -l)
+                  if [ $count -gt 9 ]; then
+                        echo "To many downloaded FoundryVTT binaries, please do $0 clean and remove at least 1 old version."
+                        break
+                  fi
+                  VERSION=$(echo "$2" | grep -oP "(?<=releases\/)\d+\.\d+")
+                  TARGET="${DEST}/${VERSION}"
 
-            FILE=$(basename "$2" | awk -F\? {'print $1'})
-            rm -rf "${DEST}/${VERSION}" >/dev/null 2>&1;
-            wget -O downloads/$FILE $2
-            echo "Extracting $FILE to ${TARGET}/ ..."
-            unzip -qq -o downloads/$FILE -d ${TARGET}/
-            VER=$(cat ${TARGET}/resources/app/package.json | jq -r '"\(.release.generation).\(.release.build)"')
-            cp ${DEST}/docker-entrypoint.sh ${TARGET}
+                  FILE=$(basename "$2" | awk -F\? {'print $1'})
+                  rm -rf "${DEST}/${VERSION}" >/dev/null 2>&1;
+                  wget -O downloads/$FILE $2
+                  echo "Extracting $FILE to ${TARGET}/ ..."
+                  unzip -qq -o downloads/$FILE -d ${TARGET}/
+                  VER=$(cat ${TARGET}/resources/app/package.json | jq -r '"\(.release.generation).\(.release.build)"')
+                  cp ${DEST}/docker-entrypoint.sh ${TARGET}
+            else
+                  echo "Usage: $0 download \"Foundry VTT Linux/NodeJS download timed URL\""
+            fi
         else
-            echo "Usage: $0 download \"Foundry VTT Linux/NodeJS download timed URL\""
+            echo "Version $MAJOR_VER not supported."
         fi
         ;;
   monitor)
